pyproject.toml
[project]
name = "par_arch_diagram"
dynamic = ["version"]
description = "AI assisted arch diagram generation tool."
url = "https://github.com/paulrobello/par_arch_diagram"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Paul Robello", email = "probello@gmail.com" }]
maintainers = [{ name = "Paul Robello", email = "probello@gmail.com" }]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Environment :: Console",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Markup :: HTML",
    "Typing :: Typed"
]
keywords = ["cloud", "diagram", "ai"]
dependencies = [
    "langchain>=0.2.16",
    "langchain-anthropic>=0.1.23",
    "langchain-community>=0.2.16",
    "langchain-core>=0.2.38",
    "langchain-experimental>=0.0.65",
    "langchain-google-genai>=1.0.10",
    "langchain-groq>=0.1.9",
    "langchain-ollama>=0.1.3",
    "langchain-openai>=0.1.23",
    "langchain-text-splitters>=0.2.4",
    "ollama>=0.3.2",
    "openai>=1.43.0",
    "pandas>=2.2.2",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.1",
    "rich>=13.8.0",
    "typer>=0.12.5",
    "botocore>=1.35.37",
    "boto3>=1.35.37",
    "langchain-aws>=0.2.2",
    "diagrams>=0.23.4",
    "langchain-xai>=0.1.1",
    "par-ai-core",
]

packages = [
    "src/par_arch_diagram"
]

[project.urls]
Homepage = "https://github.com/paulrobello/par_arch_diagram"
Documentation = "https://github.com/paulrobello/par_arch_diagram/blob/main/README.md"
Repository = "https://github.com/paulrobello/par_arch_diagram"
Issues = "https://github.com/paulrobello/par_arch_diagram/issues"
Discussions = "https://github.com/paulrobello/par_arch_diagram/discussions"
Wiki = "https://github.com/paulrobello/par_arch_diagram/wiki"

[project.scripts]
par_arch_diagram = "par_arch_diagram.__main__:app"

[build-system]
requires = ["hatchling", "wheel"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "build>=1.2.1",
    "twine>=5.1.1",
    "pyright>=1.1.379",
    "pre-commit>=3.8.0",
    "ruff>=0.7.0",
]

[tool.uv.sources]
par-ai-core = { path = "../par_ai_core" }

[tool.hatch.version]
path = "src/par_arch_diagram/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/par_arch_diagram"]
include = [
    "*.py",
    "py.typed",
    "*.png",
    "*.md",
    "*.tcss",
    "*.png",
    "*.md",
    "*.tcss"
]

[tool.hatch.build.targets.sdist]
include = [
    "src/par_arch_diagram",
    "LICENSE",
    "README.md",
    "extraction_prompt.md",
    "pyproject.toml"
]
exclude = [
    "*.pyc",
    "__pycache__",
    "*.so",
    "*.dylib"
]
